requirements.txt
fastapi
uvicorn
pydantic[dotenv]
pymongo
passlib[bcrypt]
pyjwt
